BRS-759 Part 1: extracting availability calculations to their own util
